+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add min and max date for date time element (#225)
 - Map Data Source values to elements (#224)
+- Update element Show If to support variables (#230)
 
 ## 3.1.0 (2023-08-13)